requirements.txt
asgiref==3.2.9
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-bootstrap4==2.0.1
django-multiselectfield==0.1.12
django-phonenumber-field==4.0.0
idna==2.9
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
phonenumbers==8.12.5
Pillow==7.1.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
